# docs/openapi.yaml
components:
  schemas:
...
    WharfBuildFileResponse:
      additionalProperties: false
      properties:
        createdAt:
          type: string
        id:
          format: int64
          type: integer
        size:
          format: int64
          type: integer
        state:
          enum:
            - uploading
            - uploaded
            - failed
            - evicted
            - rebuilding
          type: string
        subType:
          examples:
            - default
          type: string
        type:
          enum:
            - archive
            - patch
            - signature
          type: string
        updatedAt:
          type: string
      required:
        - id
        - size
        - state
        - type
        - subType
        - createdAt
        - updatedAt
      type: object
    WharfBuildResponse:
      additionalProperties: false
      properties:
        createdAt:
          type: string
        files:
          items:
            $ref: "#/components/schemas/WharfBuildFileResponse"
          type:
            - array
            - "null"
        id:
          format: int64
          type: integer
        parentBuild:
          $ref: "#/components/schemas/WharfParentBuildResponse"
        parentBuildId:
          format: int64
          type: integer
        state:
          enum:
            - started
            - processing
            - completed
            - failed
          type: string
        updatedAt:
          type: string
        uploadId:
          format: int64
          type: integer
        userVersion:
          type: string
        version:
          format: int64
          type: integer
      required:
        - id
        - uploadId
        - parentBuildId
        - parentBuild
        - version
        - state
        - userVersion
        - files
        - createdAt
        - updatedAt
      type: object
    WharfChannelResponse:
      additionalProperties: false
      properties:
        head:
          $ref: "#/components/schemas/WharfBuildResponse"
        name:
          type: string
        upload:
          $ref: "#/components/schemas/WharfUploadRefResponse"
      required:
        - name
        - upload
      type: object
    WharfParentBuildResponse:
      additionalProperties: false
      properties:
        id:
          format: int64
          type: integer
      required:
        - id
      type: object
    WharfUploadRefResponse:
      additionalProperties: false
      properties:
        id:
          format: int64
          type: integer
      required:
        - id
      type: object
  securitySchemes:
    ApiKeyAuth:
      description: Raw API key, Bearer token, or access_token token.
      in: header
      name: Authorization
      type: apiKey
info:
  title: DenKit Stash API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /wharf/channels:
    get:
      operationId: list-wharf-channels
      parameters:
        - description: Target in username/game form
          in: query
          name: target
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  channels:
                    additionalProperties:
                      $ref: "#/components/schemas/WharfChannelResponse"
                    type: object
                required:
                  - channels
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Not Found
      security:
        - ApiKeyAuth: []
      summary: List channels for a target game
      tags:
        - Wharf
  /wharf/channels/{channel}:
    get:
      operationId: get-wharf-channel
      parameters:
        - in: path
          name: channel
          required: true
          schema:
            type: string
        - description: Target in username/game form
          in: query
          name: target
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  channel:
                    $ref: "#/components/schemas/WharfChannelResponse"
                required:
                  - channel
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Not Found
      security:
        - ApiKeyAuth: []
      summary: Get one channel for a target game
      tags:
        - Wharf
...